update: db

+++ BillionnaireShoppingCart.Api/Models/Billionnaire.cs
@@ -3,6 +3,6 @@ namespace BillionnaireShoppingCart.Api.Models;
 public class Billionnaire : IBillionnaire
 {
     public string Name { get; set; }
-    public int NetWorth { get; set; }
+    public long NetWorth { get; set; }
     public string ImagePath { get; set; }
 }
