+++ BillionnaireShoppingCart.Api/Models/IBillionnaire.cs
@@ -3,6 +3,6 @@ namespace BillionnaireShoppingCart.Api.Models;
 public interface IBillionnaire
 {
     public string Name { get; }
-    public int NetWorth { get; }
+    public long NetWorth { get; }
     public string ImagePath { get; }
 }
